Cascaded all the changes to RepoDb.Core Solution.

+++ RepoDb.Core/RepoDb.Tests/RepoDb.IntegrationTests/Models/IdentityTable.cs
@@ -12,7 +12,7 @@ namespace RepoDb.IntegrationTests.Models
         public DateTime? ColumnDateTime { get; set; }
         public DateTime? ColumnDateTime2 { get; set; }
         public decimal? ColumnDecimal { get; set; }
-        public float? ColumnFloat { get; set; }
+        public double? ColumnFloat { get; set; }
         public int? ColumnInt { get; set; }
         public string ColumnNVarChar { get; set; }
     }
